edfi-repo-auditor/pyproject.toml
[tool.poetry]
name = "edfi-repo-auditor"
version = "0.1.0"
description = "Ed-Fi Repository Auditor"
authors = ["Ed-Fi Alliance, LLC and contributors"]
license = "Apache-2.0"
repository = "https://github.com/Ed-Fi-Exchange-OSS/DevSecOps"

[tool.poetry.dependencies]
python = "^3.9"
ConfigArgParse = "^1.5.3"
python-dotenv = "^0.20.0"
errorhandler = "^2.0.1"
requests = "^2.28.1"
pandas = "^1.4.3"
coverage = "^6.4.4"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
flake8 = "^4.0.1"
mypy = "^0.961"
pytest = "^7.1.2"
pytest-describe = "^2.0.1"
pytest-mock = "^3.8.2"
types-requests = "^2.28.1"
requests-mock = "^1.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
